+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+
+## [2.20.0] - 2023-09-18
 ### Removed
 - accountCreated, created and owned attributes to DictKey resource 
 - accountNumber and branchCode attributes to PaymentPreview resource